+++ lib/pay.rb
@@ -54,12 +54,4 @@ module Pay
   def self.setup
     yield self
   end
-
-  def self.receipts_supported?
-    charge_model.respond_to?(:receipt) &&
-      application_name.present? &&
-      business_name &&
-      business_address &&
-      support_email
-  end
 end