+++ lib/pay/stripe/webhooks/charge_refunded.rb
@@ -4,12 +4,9 @@ module Pay
       class ChargeRefunded
         def call(event)
           pay_charge = Pay::Stripe::Charge.sync(event.data.object.id, stripe_account: event.try(:account))
-          notify_user(pay_charge.customer.owner, pay_charge) if pay_charge
-        end
 
-        def notify_user(billable, charge)
-          if Pay.send_emails
-            Pay::UserMailer.with(billable: billable, charge: charge).refund.deliver_later
+          if pay_charge && Pay.send_emails
+            Pay::UserMailer.with(pay_customer: pay_charge.customer, charge: pay_charge).refund.deliver_later
           end
         end
       end