+++ lib/pay/stripe/webhooks/charge_succeeded.rb
@@ -4,12 +4,9 @@ module Pay
       class ChargeSucceeded
         def call(event)
           pay_charge = Pay::Stripe::Charge.sync(event.data.object.id, stripe_account: event.try(:account))
-          notify_user(pay_charge.customer.owner, pay_charge) if pay_charge
-        end
 
-        def notify_user(billable, charge)
-          if Pay.send_emails && charge.respond_to?(:receipt)
-            Pay::UserMailer.with(billable: billable, charge: charge).receipt.deliver_later
+          if pay_charge && Pay.send_emails
+            Pay::UserMailer.with(pay_customer: pay_charge.customer, charge: pay_charge).receipt.deliver_later
           end
         end
       end